test(editor): #304 조사권 모듈 테스트 분리
단서 조사 모듈 전용 테스트를 ModulesSubTab.deckInvestigation.test.tsx로 분리해 TSX 400줄 제한을 맞추고, window.confirm spy는 afterEach에서 항상 복구한다.

Constraint: CodeRabbit PR #340 round-2 follow-up만 처리하고 에디터 저장 계약과 runtime 동작은 변경하지 않는다.

Tested: pnpm --filter @mmp/web test -- ModulesSubTab

Tested: pnpm --filter @mmp/web typecheck

Tested: git diff --check

Related: #304
